[text-to-vector-refactoring] Moved class LanguageModelStore into an inner class of ManagedLanguageModelStore
