--- solr/modules/language-models/src/java/org/apache/solr/languagemodels/store/LanguageModelStore.java
@@ -1,66 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.solr.languagemodels.store;
-
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import org.apache.solr.languagemodels.model.SolrLanguageModel;
-
-/** Generic store to manage CRUD operations on models that extend {@link SolrLanguageModel} */
-public class LanguageModelStore<ModelT extends SolrLanguageModel> {
-
-  private final Map<String, ModelT> availableModels;
-
-  public LanguageModelStore() {
-    availableModels = Collections.synchronizedMap(new LinkedHashMap<>());
-  }
-
-  public ModelT getModel(String name) {
-    return availableModels.get(name);
-  }
-
-  public void clear() {
-    availableModels.clear();
-  }
-
-  public List<ModelT> getModels() {
-    synchronized (availableModels) {
-      final List<ModelT> availableModelsValues = new ArrayList<>(availableModels.values());
-      return Collections.unmodifiableList(availableModelsValues);
-    }
-  }
-
-  @Override
-  public String toString() {
-    return "LanguageModelStore [availableModels=" + availableModels.keySet() + "]";
-  }
-
-  public ModelT delete(String modelName) {
-    return availableModels.remove(modelName);
-  }
-
-  public void addModel(ModelT modelData) throws LanguageModelException {
-    final String name = modelData.getName();
-    if (availableModels.putIfAbsent(name, modelData) != null) {
-      throw new LanguageModelException(
-          "model '" + name + "' already exists. Please use a different name");
-    }
-  }
-}
